---Loading an old file will now correctly display the Reaper trees after load again (Reported by Kenaustin and opi-smccoole) ---"Dragon Disciple: Clear Your Mind" will now correctly add an insightful attack bonus (Reported by christhemisss) ---Parrying bonus in the following items fixed: (Reported by Texturace) ------Bladed Steel Ring ------Bracers of the Sun Soul ------Epic Swashbuckler (legacy) (level 20) ------Ethereal Bracers ------Oathblade ------Swashbuckler (legacy) (level 12) ------Swashbuckler (legacy) (level 16) ------Swashbuckler (legacy) (level 4) ------Swashbuckler (legacy) (level 8) ---Loading older file formats where "Special Feats" need to be moved from the Life to the Character object internally will no longer cause improbably stack numbers when multiple lives get loaded (Reported by DougSK) ---"Battle Engineer: Infused Weapons II" will now correctly affect repeating crossbows critical multiplier (Reported by Ogridium) ---Bad feat at Sacred Fist level 1 of "Sacred Fist(1)[/TD][TD]AC Bonus: Charisma" is now correctly "AC Bonus: Charisma" (Reported by christhemisss) ---Spell/Feat Greater Death Aura now has the correct icon ---"Deepwood Stalker: Advanced Sneak Attack" competence bonus to threat range no longer erroneously stacks with other Competence sources (Reported by christhemisss) ---Augment "Citrene of Greater Dragonmarks" is now correctly only compatible with Orange augment slots (Reported by xDTungTungTungSahur) ---Deleting the snapshot gear set will now cause breakdowns to revert to the active gear set values (Reported by xDTungTungTungSahur)
---Lamannia U74 Lamordia
------Patron "Keepers of Lamordia" added
------Patron favor feat added (Icon created)
---------Tier 1: +5% Fortification bypass
---------Tier 2: Assumed +5 HP
------Lamordia Quests added with default favor of 1
---------Locked Away (part one)
---------Zoo Creeper
---------What was lost
---------Special Delivery
---------The Body Snatchers
---------Feast and Famine
---------Ends and means
---------Locked Away (part two)
---------Body and mind
---------A Light in the Attic
---------Cold snap
---------A miner sacrifice
------New feat slot "Alter Dark Gift" added at level 8
------New Feat slot "Upgrade Dark Gift" added at level 30
------Lamordia upgrade feats added
---------Enhanced Echoing Soul
---------Enhanced Form of Pain
---------Enhanced Living Shadow
---------Enhanced Minion of the Eldritch
---------Enhanced Mist Walker
---------Enhanced Webtouched Wretch

+++ DDOBuilder/ReaperEnhancementsPane.cpp
@@ -171,7 +171,7 @@ std::vector<EnhancementTree> CReaperEnhancementsPane::DetermineTrees()
     if (pBuild != NULL)
     {
         Reaper_SelectedTrees selTrees = pBuild->ReaperSelectedTrees(); // take a copy
-        for (size_t i = 0.; i < MST_Number; ++i)
+        for (size_t i = 0.; i < static_cast<size_t>(MST_Number); ++i)
         {
             trees.push_back(GetEnhancementTree(selTrees.Tree(i)));
         }
